Add custom components
